feat(time-range-filter)!: changed format of output values, added timezone input
BREAKING CHANGE: The format of the EcTimeRangeFilterComponent's output values 'start' and 'end' was changed from numeric unix timestamps to local date (or date-time) strings according to the IANA timezone passed as a component's input (or the client's timezone if nothing passed). The time range preset functions were also affected, as well as the external time-range filter component EcTimeRangeFilterToggleComponent.